% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dataset_attachment.R
\name{add_dataset_attachment}
\alias{add_dataset_attachment}
\title{Attach any file or data to a dataset}
\usage{
add_dataset_attachment(id, attachment, description = "",
  open_browser = FALSE, curl = getCurlHandle())
}
\arguments{
\item{id}{The ID of the dataset you like to attach to. You can get the ID from the URL
of the dataset.}

\item{attachment}{Can be either a path to a file or an R data frame.}

\item{description}{An optional description to your attachmed freeformat file.}

\item{open_browser}{If this is set to true the page of the dataset is opened in the
browser after successful upload. This defaults to FALSE.}

\item{curl}{You can pass in a curl handle to reduce memory footprint and to add own options}
}
\value{
Returns a status message whether the update was successful or not.
}
\description{
This function attaches a freeformat file to a dataset.
}
\examples{
\dontrun{
rbefdata::add_dataset_attachment(id = 72, attachment = dataset, description = "This
is a test description for a new file", open_browser = T)
}
}
