\name{bef.combineObjects}
\alias{bef.combineObjects}
\title{This is a helper file that contains functions used by the BEFdata R package
Internally.}
\usage{
  bef.combineObjects(dataset_object, dataset_url,
    metadata_object, metadata_url)
}
\description{
  This is a helper file that contains functions used by the
  BEFdata R package Internally.
}

