\name{bef.getMetaData}
\alias{bef.getMetaData}
\title{Get the metadata of a dataset from the BEFdata portal}
\usage{
  bef.getMetaData(dataset_id, full_url)
}
\arguments{
  \item{dataset_id}{id of dataset in BEFdata portal.}

  \item{full_url}{direct link to its eml file.}
}
\description{
  This function fetches metedata of a dataset from the
  BEFdata portal. You need to provide the id of the
  dataset, or the direct link to its eml file.
}

