\name{bef.portal.api.dataset.metadata}
\alias{bef.portal.api.dataset.metadata}
\title{Basic api to the BEFdata portal}
\usage{
  bef.portal.api.dataset.metadata(given)
}
\arguments{
  \item{given}{Is an id of the dataset you like to get the
  information for. You find the id in the url of the
  dataset on the portal page.}
}
\value{
  Returns a list of metadata information for a dataset.
}
\description{
  This function returns the information available for a
  dataset.
}

