\name{bef.portal.api.proposal_info}
\alias{bef.portal.api.proposal_info}
\title{This function returns the information available for a paper proposal.}
\usage{
  bef.portal.api.proposal_info(id)
}
\arguments{
  \item{id}{Is the ID of the paperproposal you like to get
  information for. You can find the ID in the URL on the
  paper proposal.}
}
\value{
  Returns a list of metadata information for the proposal.
}
\description{
  This function returns the information available for a
  paper proposal.
}

