% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef.portal.get.categories.R
\name{bef.portal.get.datagroup}
\alias{bef.portal.get.datagroup}
\title{Get all or one specifc data group of interest}
\usage{
bef.portal.get.datagroup(reference, list = F)
}
\arguments{
\item{reference}{The id or name of the data groups you are interested in}

\item{list}{If this is set to true you get all datagroups as data frame}
}
\value{
Returns a data frame with the datagroup details
}
\description{
This function fetches a specific or all datagroups with all their
detail information which is a title a description and the count of
categories it contains.
}
\examples{
\dontrun{
  bef.portal.get.datagroups(list=T)
  bef.portal.get.datagroups(reference="Scientific plant speices name")
      }
}

