\name{bef.portal.get.datasets_for_keyword}
\alias{bef.portal.get.datasets_for_keyword}
\title{Fetch a list of datasets for a keyword}
\usage{
  bef.portal.get.datasets_for_keyword(keyword)
}
\arguments{
  \item{keyword}{The keyword you like to fetch the
  associated datasets for}
}
\description{
  This function fetches a list of datasets associated with
  a BEFdata portal keyword.
}
\examples{
\dontrun{
        list = bef.portal.get.datasets_for(keyword = "carbon")
      }
}

