\name{bef.pushdata}
\alias{bef.pushdata}
\title{(Unfinished) Push data to a BEFdata portal from R via the R BEFdata package}
\usage{
  bef.pushdata(dummy = "Test data pushed")
}
\arguments{
  \item{dummy}{is just a dummy parameter for the function
  this will be filled later.}
}
\value{
  It will return a message if the data was successfully
  pushed to the server.
}
\description{
  This function enables you to upload updated data and meta
  data to a BEFdata portal.
}
\examples{
\dontrun{
 message=bef.pushdata()
}
}

