\name{bef.search.tematres.keywords}
\alias{bef.search.tematres.keywords}
\title{Search a Tematres thesaurus via keywords}
\usage{
  bef.search.tematres.keywords(lookup_keyword)
}
\arguments{
  \item{lookup_keyword}{The keyword you are looking for}
}
\value{
  The function returns a vector of keywords
}
\description{
  Search a Tematres thesaurus via keywords
}

