\name{bef.searchTematres}
\alias{bef.searchTematres}
\title{Search a Tematres thesaurus via keywords}
\usage{
  bef.searchTematres(lookup_keyword)
}
\arguments{
  \item{lookup_keyword}{The keyword you are looking for}
}
\value{
  The function returns a vector of keywords
}
\description{
  Search a Tematres thesaurus via keywords
}

