\name{bef.searchTematresLower}
\alias{bef.searchTematresLower}
\title{Search a Tematres thesaurus via keywords (down hierarchy)}
\usage{
  bef.searchTematresLower(lookup_keyword)
}
\arguments{
  \item{lookup_keyword}{The keyword you are looking for}
}
\value{
  The function returns a vector of keywords
}
\description{
  Search a Tematres thesaurus via keywords (down hierarchy)
}

