\name{bef.searchTematresUpper}
\alias{bef.searchTematresUpper}
\title{Search a Tematres thesaurus via keywords (up hierarchy)}
\usage{
  bef.searchTematresUpper(lookup_keyword)
}
\arguments{
  \item{lookup_keyword}{The keyword you are looking for}
}
\value{
  The function returns a vector of keywords
}
\description{
  Search a Tematres thesaurus via keywords (up hierarchy)
}

