% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef.tematres.api.R
\name{bef.tematres.api}
\alias{bef.tematres.api}
\title{Basic tematres API via rtematres}
\usage{
bef.tematres.api(...)
}
\arguments{
\item{...}{Any task or term to be passed to the rtematres package}
}
\description{
This function queries a tematres thesaurus via the rtematres api package.
}
