% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef_options.R
\name{bef_options}
\alias{bef_options}
\title{Set or query otions related to the BEFdata R package.}
\usage{
bef_options(...)
}
\arguments{
\item{\dots}{similar to \code{\link{options}}. Note: The user credentials you
can set with the parameter 'user_credentials="xxxx"'. You find the credentials
on your profile page on the BEFdata portal that you are using.}
}
\description{
This function is used to query and set R BEFdata package specific options like
the URL to the BEFdata portal and the URL to the tematres thesaurus.
}
\examples{
# BEFdata URL
bef_options("url")
bef_options(url = "http://www.example.com")
# Tematres URL
bef_options("tematres_url")
bef_options(tematres_url = "http://www.example.com")
}
