% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datagroup.R
\name{get_datagroup}
\alias{get_datagroup}
\title{Get information of all or one specifc data group of interest}
\usage{
get_datagroup(id = NULL)
}
\arguments{
\item{id}{The ID of the data groups you are interested in.
the IDs can be found in the URL of each datagroup.}
}
\value{
Returns a data frame (tibble) with the details about the 
        datagroup you are interested in. If you leave the id parameter
        empty information about all datagruops will be shown.
}
\description{
This function fetches a specific or all datagroups on a BEFdata instance. It
provides an overview about their title, description the count of columns which
use the datagroup and the couhnt of categories inside the group.
}
\examples{
\dontrun{
bef.portal.get.datagroups(list = T)
bef.portal.get.datagroups(reference = "Scientific plant speices name")
}
}
