% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datagroup_category.R
\name{get_datagroup_category}
\alias{get_datagroup_category}
\title{Get all categories of a datagroup}
\usage{
get_datagroup_category(id)
}
\arguments{
\item{id}{The ID of the data group you like to get the categories from.}
}
\value{
Returns a data frame with the categories for the datagroup or a 
        list of data frames in case you request multiple categories.
}
\description{
This function help you get a all categories in a datagroup. This information
can also be used to harmonize your database. This can be done using the last
column in the dataframe providing IDS mentioned in the first column of the
datafram. After you have relinked the categories you can upload the data
frame using the categories upload command. The BEFdata portal will pick up
the changes and harmonize the category.
}
\examples{
\dontrun{
 get_datagroup_category(id = 22)
}

}
