% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_attachments.R
\name{get_dataset_attachment}
\alias{get_dataset_attachment}
\title{Download files attached to datasets on a BEFdata portal}
\usage{
get_dataset_attachment(id, directory = bef.options("download_dir"),
  curl = getCurlHandle())
}
\arguments{
\item{directory}{The directory to store attachment files to. By default it
creates a folder called "downloads" under the current working directory. The
default can be changed by bef.options.}

\item{dataset}{The ID of dataset you like to fetch the attachments from. You
find the ID on the dataset page on the BEFdata portal.}
}
\value{
A data frame of file information is returned invisibly. NULL is returned when
        the dataset has no attachement files.
}
\description{
This function helps you download files which are attached to a dataset on a
BEFdata instance. It downloads to a default directory in the current working
directory (downloads). This can be changed setting using the options command
of rBEFdata. If the file exists the function will choose another name with
an increasing number.
}
