% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_by_keyword.R
\name{get_dataset_by_keyword}
\alias{get_dataset_by_keyword}
\title{Fetch a list of datasets for a keyword}
\usage{
get_dataset_by_keyword(input)
}
\arguments{
\item{keyword}{The keyword you like to fetch the associated datasets for}
}
\description{
This function fetches a list of datasets associated with a BEFdata portal keyword.
}
\examples{
\dontrun{
list <- bef.portal.get.datasets.for_keyword(keyword = "carbon")
}
}
