% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata of a dataset from a BEFdata portal instance}
\usage{
get_metadata(id, split_category = T)
}
\arguments{
\item{split_category}{Determines whether the metadata of a column with mixed
data types is split into two separate numeric and factorial fields to
describe the columns. See also split_category for the dataset download
function.}

\item{dataset}{This is either a dataset ID or a file path to a eml metadata
file.}
}
\value{
A list of metadata. Metadata That doesn't exist is represented as \code{NA}
}
\description{
This function reads and imports metadata of a dataset on a BEFdata portal.
You need to provide the function with the IDs of the datasets. You can find
of a dataset in the URL of the dataset page in the portal.
}
