% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposal_dataset.R
\name{get_proposal_dataset}
\alias{get_proposal_dataset}
\title{Fetch primary from a BEFdata portal paperproposal}
\usage{
get_proposal_dataset(id, split_category = T)
}
\arguments{
\item{id}{This is the ID of the paper proposal you want to get}

\item{split_category}{Determines whether columns with mixed data types are split
into two separate numeric and factorial columns}
}
\value{
The function returns a list of the datasetsattached to a proposal.
}
\description{
This function fetches data from a paperproposal on a BEFdata portal. This
allows a fast import of your data into R to start with your analysis. You
need to provide the function with a proposal id which you can find in the
URL of the proposal.
}
\examples{
\dontrun{
     prop1 <- get_proposal_dataset(id = 8)
}
}
