% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbefdata_defunct.R
\name{rbefdata_defunct}
\alias{rbefdata_defunct}
\alias{bef.getDataset}
\alias{bef.getKeywords}
\alias{bef.getMetadata}
\alias{bef.getProposal}
\alias{bef.portal.get}
\alias{bef.tematres.api}
\alias{bef.mediawiki.api.define}
\alias{bef.portal.api.dataset_info}
\alias{bef.portal.api.proposal_info}
\alias{bef.tematres}
\title{Defunct functions in package rbefdata}
\usage{
bef.getDataset(...)

bef.getKeywords()

bef.getMetadata(...)

bef.getProposal(...)

bef.portal.get(...)

bef.tematres.api(...)

bef.tematres(...)

bef.mediawiki.api.define(...)

bef.portal.api.dataset_info(...)

bef.portal.api.proposal_info(...)
}
\description{
These following functions are renamed or removed. Please use new functions instead.
}
\details{
\tabular{ll}{
 \strong{Defunct functions} \tab \strong{Replacement} \cr
 \code{bef.getDataset} \tab \code{\link{bef.portal.get.dataset}} \cr
 \code{bef.getMetadata} \tab \code{\link{bef.portal.get.metadata}} \cr
 \code{bef.getKeywords} \tab \code{\link{bef.portal.get.keywords}} \cr
 \code{bef.getProposal} \tab \code{\link{bef.portal.get.proposal}} \cr
 \code{bef.portal.get} \tab no replacement \cr
 \code{bef.tematres.api} \tab no replacement \cr
 \code{bef.tematres} \tab no replacement \cr
 \code{bef.mediawiki.api.define} \tab no replacement \cr
 \code{bef.portal.api.dataset_info} \tab no replacement \cr
 \code{bef.portal.api.proposal_info} \tab no replacement \cr
}
}
\keyword{internal}
