\name{bef.combineObjects}
\alias{bef.combineObjects}
\title{This is a helper file that contains functions used by the BEFdata R package
Internally.}
\usage{
  bef.combineObjects(dataset_object, dataset_url,
    metadata_object, metadata_url)
}
\arguments{
  \item{dataset_object}{This a BEFdata dataset}

  \item{dataset_url}{This is the url to a BEFdata dataset}

  \item{metadata_object}{This is a metadata object you can
  get via bef.getMetadata}

  \item{metadata_url}{This is the URL to a BEFdata dataset
  eml file}
}
\description{
  This is a helper file that contains functions used by the
  BEFdata R package Internally.
}

