\name{bef.getMetadata}
\alias{bef.getMetadata}
\title{Get the metadata of a dataset from the BEFdata portal}
\usage{
  bef.getMetadata(dataset_id, full_url)
}
\arguments{
  \item{dataset_id}{The id of the dataset in the BEFdata
  portal.}

  \item{full_url}{Functions as direct download link for the
  eml file.}
}
\description{
  This function fetches the metadata associated with
  dataset on a BEFdata portal. You need to provide the id
  of the dataset, or the direct link to the Ecological
  Metadata Language file. You can find the url on the
  dataset page.
}

