\name{bef.portal.get.keywords}
\alias{bef.portal.get.keywords}
\title{Fetch keywords from a BEFdata portal.}
\usage{
  bef.portal.get.keywords()
}
\value{
  The function returns a character vector of keywords.
}
\description{
  This function fetches keywords from a BEFdata portal.
}
\examples{
\dontrun{
            keywords=bef.portal.get.keywords()
          }
}

