% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bef.portal.get.metadata.R
\name{bef.portal.get.metadata}
\alias{bef.portal.get.metadata}
\alias{bef.get.metadata}
\alias{bef.get.metadata_for}
\alias{bef.portal.get.metadata_for}
\title{Get the metadata of a dataset from the BEFdata portal}
\usage{
bef.portal.get.metadata(dataset, split_category = T)
}
\arguments{
\item{dataset}{This is either a dataset ID or a file path to a eml metadata file.}

\item{split_category}{Determines whether the metadata of a column with mixed data types
is split into two separate numeric and factorial fields to describe the columns.
See also split_category for the dataset download function.}
}
\value{
A list of metadata. Metadata That doesn't exist is represented as \code{NA}
}
\description{
This function fetches the metadata associated with a dataset on a BEFdata portal.
You need to provide the function with an id of the dataset or a path to an eml
file you downloaded from the portal manually. You can find the ID in the URL of
the dataset page in the portal (e.g ID 6: .../datasets/6).
}
