% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bef.portal.upload.R
\name{bef.portal.upload.categories}
\alias{bef.portal.upload.categories}
\alias{bef.upload.categories}
\title{Upload a category set}
\usage{
bef.portal.upload.categories(datagroup_id, categories, curl = getCurlHandle())
}
\arguments{
\item{datagroup_id}{Categories ID you like to harmonize. You can find this in the URL of
the category.}

\item{categories}{Either a data frame with the categories or a csv containing the categories
and merge ID and stuff downloaded from the portal.}

\item{curl}{Pass in a curl handle with own options or to reduce memory footprint.}
}
\value{
Returns a status message whether the upload was successful or not.
}
\description{
This function changes the categories of a datagroup
on a BEFdata portal. It expects categories in the exact
same format as they are downloaded using the
"bef.portal.get.categories_for" command. These are five
columns specifying the category id, the short, long, and
description field of the category, as well as a "merge id"
column. The latter is used if a category should be merged
with another one. The category given in teh "merge id"
column will stay on the portal.
}
\examples{
\dontrun{
  dataframe = bef.portal.get.categories_for(datagroup_id = 22)
  bef.portal.upload.categories(datagroup_id = 22, categories = dataframe)
      }
}
\seealso{
\code{bef.portal.get.categories_for}
}

