\name{rbefdata-deprecated}
\alias{bef.tematres.search.broader_keywords}
\alias{bef.tematres.search.keywords}
\alias{bef.tematres.search.narrower_keywords}
\alias{rbefdata-deprecated}
\title{Deprecated functions in package rbefdata}
\usage{
  bef.tematres.search.keywords(...)

  bef.tematres.search.narrower_keywords(...)

  bef.tematres.search.broader_keywords(...)
}
\description{
  These following functions are renamed or removed. Please
  use new functions instead.
}
\details{
  \tabular{ll}{ \strong{Deprecated functions} \tab
  \strong{Replacement} \cr
  \code{bef.tematres.search.keywords} \tab
  \code{\link{bef.tematres.api.search}} \cr
  \code{bef.tematres.search.narrower_keywords} \tab
  \code{\link{bef.tematres.api.search}} \cr
  \code{bef.tematres.search.broader_keywords} \tab
  \code{\link{bef.tematres.api.search}} \cr }
}
\keyword{internal}

