\name{rbefdata-deprecated}
\alias{bef.portal.get.dataset_list}
\alias{bef.portal.get.proposal}
\alias{bef.tematres.api}
\alias{rbefdata-deprecated}
\title{Deprecated functions in package rbefdata}
\usage{
bef.portal.get.dataset_list(...)

bef.portal.get.proposal(...)

bef.tematres.api(...)
}
\description{
These following functions are renamed or removed. Please
use new functions instead.
}
\details{
\tabular{ll}{ \strong{Deprecated functions} \tab
\strong{Replacement} \cr \code{bef.portal.get.dataset_list}
\tab \code{\link{bef.portal.get.datasets.for_keyword}} \cr
\code{bef.portal.get.proposal} \tab
\code{\link{bef.portal.get.datasets.for_proposal}} \cr
\code{bef.tematres.api} \tab \code{\link{bef.tematres}} \cr
}
}
\keyword{internal}

