% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbefdata-deprecated.R
\name{rbefdata-deprecated}
\alias{rbefdata-deprecated}
\alias{bef.portal.get.dataset_list}
\alias{bef.portal.get.proposal}
\alias{bef.tematres.api}
\alias{bef.get.dataset}
\alias{bef.get.dataset_by}
\alias{bef.portal.get.dataset_by}
\alias{bef.portal.get.dataset}
\alias{bef.get.metadata}
\alias{bef.get.metadata_for}
\alias{bef.portal.get.metadata}
\alias{bef.portal.get.metadata_for}
\alias{bef.get.attachments}
\alias{bef.get.attachments_for}
\alias{bef.portal.get.attachments}
\alias{bef.portal.get.attachments_for}
\alias{bef.portal.datagroup}
\alias{bef.portal.get.datagroup}
\title{Deprecated functions in package rbefdata}
\usage{
bef.portal.get.dataset_list(...)

bef.portal.get.proposal(...)

bef.tematres.api(...)

bef.get.dataset(...)

bef.get.dataset_by(...)

bef.portal.get.dataset_by(...)

bef.portal.get.dataset(...)

bef.get.metadata(...)

bef.get.metadata_for(...)

bef.portal.get.metadata(...)

bef.portal.get.metadata_for(...)

bef.get.attachments(...)

bef.get.attachments_for(...)

bef.portal.get.attachments(...)

bef.portal.get.attachments_for(...)

bef.portal.get.datagroup(...)

bef.portal.datagroup(...)
}
\description{
These following functions are renamed or removed. Please use new functions instead.
}
\details{
\tabular{ll}{
 \strong{Deprecated functions} \tab \strong{Replacement} \cr
 \code{bef.portal.get.dataset_list} \tab \code{\link{bef.portal.get.datasets.for_keyword}} \cr
 \code{bef.portal.get.proposal} \tab \code{\link{bef.portal.get.datasets.for_proposal}} \cr
 \code{bef.tematres.api} \tab \code{\link{bef.tematres}} \cr
 \code{bef.get.dataset} \tab \code{\link{get_dataset}} \cr
 \code{bef.get.dataset_by} \tab \code{\link{get_dataset}} \cr
 \code{bef.portal.get.dataset_by} \tab \code{\link{get_dataset}} \cr
 \code{bef.portal.get.dataset} \tab \code{\link{get_dataset}} \cr
 \code{bef.get.metadata} \tab \code{\link{get_metadata}} \cr
 \code{bef.get.metadata_for} \tab \code{\link{get_metadata}} \cr
 \code{bef.portal.get.metadata} \tab \code{\link{get_metadata}} \cr
 \code{bef.portal.get.metadata_for} \tab \code{\link{get_metadata}} \cr
 \code{bef.get.attachments} \tab \code{\link{get_dataset_attachment}} \cr
 \code{bef.get.attachments_for} \tab \code{\link{get_dataset_attachment}} \cr
 \code{bef.portal.get.attachments} \tab \code{\link{get_dataset_attachment}} \cr
 \code{bef.portal.get.attachments_for} \tab \code{\link{get_dataset_attachment}} \cr
 \code{bef.portal.datagroup} \tab \code{\link{get_datagroup}} \cr
 \code{bef.portal.get.datagroup} \tab \code{\link{get_datagroup}} \cr
}
bef.portal.get.dataset_list
}
\keyword{internal}
