\name{rbefdata-deprecated}
\alias{bef.portal.get.dataset_list}
\alias{bef.portal.get.proposal}
\alias{rbefdata-deprecated}
\title{Deprecated functions in package rbefdata}
\usage{
  bef.portal.get.dataset_list(...)

  bef.portal.get.proposal(...)
}
\description{
  These following functions are renamed or removed. Please
  use new functions instead.
}
\details{
  \tabular{ll}{ \strong{Deprecated functions} \tab
  \strong{Replacement} \cr
  \code{bef.portal.get.dataset_list} \tab
  \code{\link{bef.portal.get.datasets.for_keyword}} \cr
  \code{bef.portal.get.proposal} \tab
  \code{\link{bef.portal.get.datasets.for_proposal}} \cr }
}
\keyword{internal}

