% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbefdata-package.R
\docType{package}
\name{rbefdata-package}
\alias{rbefdata}
\alias{rbefdata-package}
\title{rbefdata: BEFdata R Package}
\description{
Basic R package to access data structures offered by any 'BEFdata' portal instance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ataeub/rbefdata}
}

}
\author{
\strong{Maintainer}: Alexander Täuber \email{alexander.taeuber@idiv.de}

Authors:
\itemize{
  \item Claas-Thido Pfaff \email{ctpfaff@arcor.de}
  \item Karin Nadrowski \email{nadrowski@uni-leipzig.de}
  \item Xingxing Man \email{manxingxing@gmail.com}
}

}
\keyword{internal}
