\docType{package}
\name{rbefdata-package}
\alias{rbefdata}
\alias{rbefdata-package}
\title{A package to fetch datasets and metadata information from a BEFdata portal}
\description{
  The package is by default associated to the websites of
  the BEF-China experiment:
  \url{http://befdataproduction.biow.uni-leipzig.de}.
}
\details{
  If you use the package with your own instance of a
  BEFdata portal you need to change the default URLs to
  your needs via bef.options().
}

